Add register value poll function to library
Takes two arguments, the first is the register that will be polled and the 2nd
is the value the register must be equal to or greater than before the function
returns. The function will wait forever if the value of the register is not
equal to or greater then the 'poll' value.